refactor: extract nested structs into separate files
Move FlowLayout, ServerStatusView, ScrollWheelRedirector, and
ScrollWheelMonitorView into their own files to reduce view file size.
FolderRow kept inline in SourceListView (tiny, tightly coupled).
